requirements.txt
streamlit
rdkit-pypi
pandas
view.setBackgroundColor('white')